fix bug when instantiate NovaImage in NovaServer
After i boot a server from a cinder volume in icehouse . I want to get
the server instance with NovaServer Object, but it throw a exception
because openstack return server json where value of imgae is  empty
string, not null. Class NovaServer's imgae is instance of NovaImage,
instantiate failed and throws an exception. So I change the NovaImage to
Object, and modify getImgae() to satisfy.