Respect original jar names.
This patch uses maven-jar-plugin to generate openstack4j project jars,
just as it was done before introducing maven-bundle-plugin.
This way, previously used instructions are still used resulting in the same jar names.

In order to create OSGi comaptible jars (bundles),
this patch adds to maven-jar-plugin instructions to attach the bundle manifest file to generated jar.
maven-bundle-plugin is used with the solely purpose of generating the bundle manifest file.