DLPayload returns response to allow proper handling
Currently, if getting an object is not successful, the user does not
know whether the GET was successful or not.  For example, if the result
is a 404, then reading from the InputStream will result in the stream
containing the message saying data not available instead of the data
that is actually wanted.

Steps to reproduce:
Try to download a file that does not exist. Reading the input stream

Proposed solution:
Allow users to check the response to see whether the request is
successful or not before reading form the input stream.