Adding new ProjectService API for listing projects
Adding new API for projects listing on certain domain
in ProjectService.

Keystone expects domain_id value in URL for domain scoped
authentication tokens to be authorized to list domain's projects.

OS4J ProjectService is not aware of the auth. token scope domain.

This patch allows to use projects list API
in a domain scope by specifying the domai_id as a parameter
in the keystone projects URL.

Link to an issue: #1122